Added tests for restarting docker service.
